adding implementing laravel

+++ LaravelApplicationDevelopmentBlueprints/l4-e-commerce-website/app/database/migrations/2014_02_10_230138_create_books_table.php
@@ -15,7 +15,7 @@ class CreateBooksTable extends Migration {
 		Schema::create('books', function(Blueprint $table) {
 			$table->increments('id');
 			$table->string('title');
-			$table->integer('isbn');
+			$table->decimal('isbn', 15);
 			$table->string('cover');
 			$table->decimal('price', 10, 2);
 			$table->integer('author_id')->unsigned()->default(0);
